+++ app/router/middleware.go
@@ -1,4 +1,8 @@
 package router
 
 
-//TODO create middleware
+//TODO create basic middleware wrapper
+// example MiddleWareHandler(next Handler){
+	// autheticate
+	// next.ServeHTTP()
+// }
