+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/lucas-kern/tower-of-babel_server
 
 go 1.18
+
+require github.com/julienschmidt/httprouter v1.3.0 // indirect
